# docs/api/openapi.yaml
components:
  schemas:
    ActivityLogOut:
      properties:
        actor_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Actor Id
        at:
          format: date-time
          title: At
          type: string
        entity:
          anyOf:
          - type: string
          - type: 'null'
          title: Entity
        entity_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Entity Id
        event:
          title: Event
          type: string
        id:
          format: uuid
          title: Id
          type: string
        ip:
          anyOf:
          - type: string
          - type: 'null'
          title: Ip
        payload:
          anyOf:
          - type: object
          - type: 'null'
          title: Payload
      required:
      - id
      - event
      - at
      title: ActivityLogOut
      type: object
    AuditLogOut:
      properties:
        action:
          title: Action
          type: string
        actor_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Actor Id
        actor_role:
          anyOf:
          - type: string
          - type: 'null'
          title: Actor Role
        after:
          anyOf:
          - type: object
          - type: 'null'
          title: After
        at:
          format: date-time
          title: At
          type: string
        before:
          anyOf:
          - type: object
          - type: 'null'
          title: Before
        diff:
          anyOf:
          - type: object
          - type: 'null'
          title: Diff
        entity:
          title: Entity
          type: string
        entity_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Entity Id
        id:
          format: uuid
          title: Id
          type: string
        ip:
          anyOf:
          - type: string
          - type: 'null'
          title: Ip
        request_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Request Id
        user_agent:
          anyOf:
          - type: string
          - type: 'null'
          title: User Agent
      required:
      - id
      - entity
      - action
      - at
      title: AuditLogOut
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    InvitationCreateIn:
      properties:
        email:
          format: email
          title: Email
          type: string
        role_code:
          title: Role Code
          type: string
      required:
      - email
      - role_code
      title: InvitationCreateIn
      type: object
    InvitationOut:
      properties:
        accepted_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Accepted At
        created_at:
          format: date-time
          title: Created At
          type: string
        email:
          format: email
          title: Email
          type: string
        expires_at:
          format: date-time
          title: Expires At
          type: string
        id:
          format: uuid
          title: Id
          type: string
        role_id:
          format: uuid
          title: Role Id
          type: string
        token:
          title: Token
          type: string
      required:
      - id
      - email
      - role_id
      - token
      - expires_at
      - created_at
      title: InvitationOut
      type: object
    MeOut:
      properties:
        email:
          format: email
          title: Email
          type: string
        full_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Full Name
        id:
          format: uuid
          title: Id
          type: string
        last_login_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Last Login At
        permissions:
          items:
            type: string
          title: Permissions
          type: array
        roles:
          items:
            type: string
          title: Roles
          type: array
        status:
          title: Status
          type: string
      required:
      - id
      - email
      - status
      title: MeOut
      type: object
    Page_ActivityLogOut_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ActivityLogOut'
          title: Items
          type: array
        page:
          title: Page
          type: integer
        page_size:
          title: Page Size
          type: integer
        total:
          title: Total
          type: integer
      required:
      - items
      - page
      - page_size
      - total
      title: Page[ActivityLogOut]
      type: object
    Page_AuditLogOut_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/AuditLogOut'
          title: Items
          type: array
        page:
          title: Page
          type: integer
        page_size:
          title: Page Size
          type: integer
        total:
          title: Total
          type: integer
      required:
      - items
      - page
      - page_size
      - total
      title: Page[AuditLogOut]
      type: object
...
    SessionOut:
      properties:
        created_at:
          format: date-time
          title: Created At
          type: string
        expires_at:
          format: date-time
          title: Expires At
          type: string
        id:
          format: uuid
          title: Id
          type: string
        ip:
          anyOf:
          - type: string
          - type: 'null'
          title: Ip
        jti:
          title: Jti
          type: string
        revoked_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Revoked At
        user_agent:
          anyOf:
          - type: string
          - type: 'null'
          title: User Agent
      required:
      - id
      - jti
      - created_at
      - expires_at
      title: SessionOut
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  title: Kaya BOS
  version: 1.0.0
openapi: 3.1.0
paths:
  /api/v1/audit/activity:
    get:
      operationId: list_activity_api_v1_audit_activity_get
      parameters:
      - in: query
        name: page
        required: false
        schema:
          default: 1
          minimum: 1
          title: Page
          type: integer
      - in: query
        name: page_size
        required: false
        schema:
          default: 50
          maximum: 200
          minimum: 1
          title: Page Size
          type: integer
      - in: query
        name: event
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Event
      - in: query
        name: actor_id
        required: false
        schema:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Actor Id
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_ActivityLogOut_'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Activity
      tags:
      - audit
  /api/v1/audit/logs:
    get:
      operationId: list_audit_api_v1_audit_logs_get
      parameters:
      - in: query
        name: page
        required: false
        schema:
          default: 1
          minimum: 1
          title: Page
          type: integer
      - in: query
        name: page_size
        required: false
        schema:
          default: 50
          maximum: 200
          minimum: 1
          title: Page Size
          type: integer
      - in: query
        name: entity
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Entity
      - in: query
        name: entity_id
        required: false
        schema:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Entity Id
      - in: query
        name: actor_id
        required: false
        schema:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Actor Id
      - in: query
        name: action
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Action
      - in: query
        name: since
        required: false
        schema:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Since
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_AuditLogOut_'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Audit
      tags:
      - audit
  /api/v1/auth/logout:
    post:
      operationId: logout_api_v1_auth_logout_post
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Logout
      tags:
      - auth
  /api/v1/auth/me:
    get:
      operationId: me_api_v1_auth_me_get
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Me
      tags:
      - auth
  /api/v1/auth/sessions:
    get:
      operationId: list_sessions_api_v1_auth_sessions_get
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/SessionOut'
                title: Response List Sessions Api V1 Auth Sessions Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Sessions
      tags:
      - auth
...
  /api/v1/invitations:
    get:
      operationId: list_invitations_api_v1_invitations_get
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/InvitationOut'
                title: Response List Invitations Api V1 Invitations Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Invitations
      tags:
      - invitations
    post:
      operationId: create_invitation_api_v1_invitations_post
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvitationCreateIn'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Invitation
      tags:
      - invitations
  /api/v1/invitations/{invitation_id}:
    delete:
      operationId: revoke_invitation_api_v1_invitations__invitation_id__delete
      parameters:
      - in: path
        name: invitation_id
        required: true
        schema:
          format: uuid
          title: Invitation Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Revoke Invitation
      tags:
      - invitations
...